% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_sql.R
\name{assemble_sql_tbls}
\alias{assemble_sql_tbls}
\title{Assemble results stored in SQLite data base to a tibble}
\usage{
assemble_sql_tbls(sql_path, sql_name)
}
\arguments{
\item{sql_path}{Path of the SQLite data base as a character string}
}
\description{
Analyses generated in CapRosion are often organized in chunks of tibbles
stored in SQLite data bases. This function helbs to load the tibble chunks
from a data base and to assemble them to one results table
}
