% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_raster.R
\name{format_raster}
\alias{format_raster}
\title{Format raster before returning}
\usage{
format_raster(rst)
}
\arguments{
\item{rst}{Raster layer}
}
\description{
Format raster before returning
}
\keyword{internal}
