% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_raster.R
\name{make_raster_from_tibble}
\alias{make_raster_from_tibble}
\title{Create raster layers from tibbles}
\usage{
make_raster_from_tibble(tbl, ref_raster, ref_idx, variable = NULL,
  write_path = NULL, and_return = FALSE)
}
\arguments{
\item{tbl}{Result tibble}

\item{ref_raster}{The reference mask raster layer (Stored as 'mask_ugke.tif'
on the hard drive). \code{ref_raster} has to be the loaded raster layer or
the full path to the '.tif' on the hard drive.}

\item{ref_idx}{Logical vector that defines the raster cells of the raster
layer to which result values should be assigned (Stored as
'has_value_mask.rds' on the hard drive). \code{ref_idx} has to be the
loaded index vector or the full path to the '.rds' on the hard drive.}

\item{variable}{(Optional) Character vector to define the variables in the
tibble that should be converted to a raster layer. If \code{variable} is
not defined, all columns are converted}

\item{write_path}{(Optional) Character string to defie the path where the
raster layers should be written as GeoTiffs. If \code{write_path} is not
defined, the raster layers are returned to R by default.}

\item{and_return}{(Optional) Logical. If \code{and_return = TRUE}, the raster
layers are additionally returned in R, when they are written to the hard
drive (write_path is set). drive}
}
\description{
Analysis results loaded in tibble format can be overlayed with the project's
raster mask layer with this function
}
