% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_raster.R
\name{pass_raster_value}
\alias{pass_raster_value}
\title{Pass values to raster and assign at the right positions}
\usage{
pass_raster_value(rst, val, pos)
}
\arguments{
\item{rst}{Raster layer}

\item{val}{Values vector}

\item{pos}{Position vector}
}
\description{
Pass values to raster and assign at the right positions
}
\keyword{internal}
