% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_sql.R
\name{sort_by_index}
\alias{sort_by_index}
\title{Sort names by their suffix number}
\usage{
sort_by_index(lbls)
}
\arguments{
\item{lbls}{Vector of character strings with numbers in name that should be
sorted}
}
\description{
Sort names by their suffix number
}
\keyword{internal}
